# docker: Speed up startup of the materialized container