--- a/src/materialized/ci/entrypoint.sh
+++ b/src/materialized/ci/entrypoint.sh
@@ -41,15 +41,18 @@ EOF
 
 # Start PostgreSQL, unless suppressed.
 if [ -z "${MZ_NO_BUILTIN_POSTGRES:-}" ]; then
+  (trap 'pg_ctlcluster 16 main stop --mode=fast --force' SIGTERM SIGINT
+  rm -f /var/run/postgresql/.s.PGSQL.26257.lock
   pg_ctlcluster 16 main start
-  psql -U root -c "CREATE SCHEMA IF NOT EXISTS consensus"
-  psql -U root -c "CREATE SCHEMA IF NOT EXISTS storage"
-  psql -U root -c "CREATE SCHEMA IF NOT EXISTS adapter"
-  psql -U root -c "CREATE SCHEMA IF NOT EXISTS tsoracle"
+  psql -U postgres -c "CREATE ROLE root WITH LOGIN PASSWORD 'root'" || true
+  psql -U postgres -c "CREATE DATABASE root OWNER root" || true
+  psql -U root -c "CREATE SCHEMA IF NOT EXISTS consensus; CREATE SCHEMA IF NOT EXISTS storage; CREATE SCHEMA IF NOT EXISTS adapter; CREATE SCHEMA IF NOT EXISTS tsoracle") &
 fi
 
 # Start nginx to serve the console.
-nginx
+if [ -z "${MZ_NO_BUILTIN_CONSOLE:-}" ]; then
+  nginx &
+fi
 
 if [[ ! -f /mzdata/environment-id ]]; then
   echo "docker-container-$(cat /proc/sys/kernel/random/uuid)-0" > /mzdata/environment-id